# artifacts/api-server/src/routes/telegram.ts
import { Router, type IRouter } from "express";
import { Bot } from "grammy";
import { db } from "@workspace/db";
import { telegramUsersTable, messagesTable } from "@workspace/db";
import { eq } from "drizzle-orm";
import { logger } from "../lib/logger";

const router: IRouter = Router();

const BOT_TOKEN = process.env["TELEGRAM_BOT_TOKEN"];

let bot: Bot | null = null;

if (BOT_TOKEN) {
  bot = new Bot(BOT_TOKEN);

  bot.command("start", async (ctx) => {
    await ctx.reply(
      "¡Hola! Soy el bot del proyecto Buildathon. Escríbeme algo y lo guardaré.",
    );
  });

  bot.on("message:text", async (ctx) => {
    const from = ctx.message.from;
    const text = ctx.message.text;

    try {
      let user = await db.query.telegramUsersTable.findFirst({
        where: eq(telegramUsersTable.telegramId, from.id),
      });

      if (!user) {
        const [created] = await db
          .insert(telegramUsersTable)
          .values({
            telegramId: from.id,
            username: from.username ?? null,
            firstName: from.first_name,
            lastName: from.last_name ?? null,
          })
          .returning();
        user = created;
      }

      if (user) {
        await db.insert(messagesTable).values({
          telegramMessageId: ctx.message.message_id,
          userId: user.id,
          text,
          direction: "incoming",
        });

        const reply = `Recibí tu mensaje: "${text}"`;
        await ctx.reply(reply);

        await db.insert(messagesTable).values({
          telegramMessageId: ctx.message.message_id + 1,
          userId: user.id,
          text: reply,
          direction: "outgoing",
        });
      }
    } catch (err) {
      logger.error({ err }, "Error handling Telegram message");
    }
  });
}

router.post("/telegram/webhook", async (req, res) => {
  if (!bot) {
    res.status(503).json({ ok: false, error: "Bot not configured. Set TELEGRAM_BOT_TOKEN." });
    return;
  }

  try {
    await bot.handleUpdate(req.body);
    res.json({ ok: true });
  } catch (err) {
    req.log.error({ err }, "Webhook error");
    res.status(500).json({ ok: false });
  }
});

export { bot };
export default router;