# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Telegram Chatbot + PWA
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WebhookResponse {
  ok: boolean;
}

export interface TelegramUser {
  id: number;
  telegramId: number;
  /** @nullable */
  username?: string | null;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  createdAt: string;
}

export type MessageDirection =
  (typeof MessageDirection)[keyof typeof MessageDirection];

export const MessageDirection = {
  incoming: "incoming",
  outgoing: "outgoing",
} as const;

export interface Message {
  id: number;
  telegramMessageId: number;
  userId: number;
  /** @nullable */
  text?: string | null;
  direction: MessageDirection;
  createdAt: string;
  user: TelegramUser;
}

export interface MessagesResponse {
  messages: Message[];
  total: number;
}

export interface Stats {
  totalMessages: number;
  totalUsers: number;
  messagesLast24h: number;
  activeUsersLast24h: number;
}

export type TelegramWebhookBody = { [key: string]: unknown };

export type GetMessagesParams = {
  limit?: number;
  offset?: number;
  /**
   * @nullable
   */
  userId?: number | null;
};