# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Telegram Chatbot + PWA
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /telegram/webhook:
    post:
      operationId: telegramWebhook
      tags: [messages]
      summary: Telegram webhook endpoint
      description: Receives incoming updates from Telegram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
...
components:
  schemas:
...
    WebhookResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
...